moved out extension of UIImageView to separate file to have OS X support for DownloadableImage. Also changed UIImage to Image typealias
