+++ RxExample/RxExample/Services/ReachabilityService.swift
@@ -6,7 +6,9 @@
 //  Copyright © 2015 Krunoslav Zaher. All rights reserved.
 //
 
+#if !RX_NO_MODULE
 import RxSwift
+#endif
 
 public enum ReachabilityStatus{
     case Reachable,Unreachable
